INTERIM more or less figured out crashularity
SOOOOO... pointers to member functions.  in order to support
virtual functions, member function pointers are actually
pointer + adjustment to "this" so that "this" can point
to the vtable for the appropriate subclass of whatever the
this is.  this appears to be boobytrap-esque.

Going to either make states static members or wrap states
in a class or ... something.