rearrange relationship between "base" parser and generated
"Base" parser template now takes only an implementation class.
The implementation class (the one we generate) contains a
member instance for the "base" parser.  This removes some
chicken-and-egg issues an should let us reduce the implementation
class to just the set of tokens and states, allowing us to
keep all (or most of) the "real" code in the base parser.

Also use the chicken-egg reduction advantage to make it so
we can print the names of nonterminals, instead of just the
numeric IDs, which makes debug output much easier to understand.