interim - warn if more than one optional per state
So for the moment I'm going to punt the thorny issue of which
thing to shift if there's more than one optional in a given state.
This is thorny because we need to shift a placeholder for optionals,
but we don't necessarily know which next state is applicable if
there's more than one optional item.  i.e. we don't know if we should
shift the nothing for optional a or for optional b.

There are definitely cases where it's not actually ambiguous, but
to simplify my life I'm going to put the onus of avoiding this sort
of ambiguity on the fpl authors.