code.format $LINE restoration optional; newlines after $LINE
Make it so that line restoration is optional in code formatting,
so that we can change the reduce code handing such that if (eg)
the fpl author forgets a "return" statement, the line number
for that warning is reported as coming from a line which is at
least somewhat close to the line in the fpl source.  (it'll
actually be the real line + 1, because the end of the function
is inserted by fpl2cc.  good enough for the moment).

Also make sure there's a newline after a #$LINE restoration
so that we don't accidentally jam other code onto the #line
preprocessor line.