Repetition works at all
Make repetition work.  To that end, add the concept of the
"stack slice", and pass those to reduction rules for things
with not-guaranteed-single repetitions.  Internally, other
arguments are also implemented via stack slices, but they
are passed as Products or std::strings as before.

repetition test fpl tests itself now.  this might be the
direction to take with fpl tests, for now anyway, though
it's not great because it means the .test depends on stuff
embedded in the .fpl.  Perhaps if I were smart, I'd set
up a general way to embed testing code in the .test files.
Not now though.  plowing on.