fpl2cc: fix comment line number; try freidnlier defaults
Fix the __LINE__ cludge in fpl2cc so that it actually shows
the right (for some definition) line number.

Try to make the default rules more likely to compile and less
likely to make code which blows up with inscrutable error
messages.

.. really need to work out some kind of default product type.
Also probably want to support different product types...