fpl SConstruct fail dependency; repetition test update
Make scons aware of test failures (by making it evaluate to false)
so that we get "build failed" or whatever output if tests fail.

Update the repetition.fpl and test, basically just noodling because
something odd is happening in another fpl I have and thought I
could reproduce the problem (but could not).  hmm