fix missing includes, const in vector (!?), misc
Trying to get this to compile under g++ on ubuntu, I came across some
missing headers and some other things gnu complained about.

The missing headers are fixed and are definitely the right thing.

There was a warning about not using a return value in a write() call,
which I think is a silly warning, but I fixed it anyway.

Last, and most weirdly and importantly, YOU CAN'T HAVE CONST MEMBERS
OF AN ARRAY in gnu c++.  Wat?!?!?!?  Turns out it's (probably?)
according to standard!  The c++ standard is a bloated ridiculous
nightmare.

This version (probably) doesn't work under g++ still;  I'm merging
changes which I think will be (relatively) uncontrovertial first.

added missing #include<cassert>

.. hates the assert.

maybe this should actually be a static assert anyway.  punting
for the moment as I try to get this to compile on gcc

add missing #include "limits.h"

No const in vectors?!?!!

To my surprise, under gnu, you can't have a vector of const objects!

To my even greater surprise, this behaviour is "correct" under
the c++ standard (at least according to stack overflow):

https://stackoverflow.com/questions/6954906/does-c11-allow-vectorconst-t

The c++ 17 standard is a mere 1600 or so pages (pdf so as to be hard to
search) so it should be no problem to just flip through and find it.
In fact I can't believe anyone wouldn;t just ahve all this in their
heads:

http://www.open-std.org/jtc1/sc22/wg21/docs/papers/2017/n4659.pdf

anyway uhh.. the code still seems to work.

json.fpl shows total bytes in binary version

I did this to shut up a warning it gets under gnu, but I find
I actually do want to know how big it is vs the source