reasonable error message on regex compile failure
When regexes fail to compile, the default exception thrown
doesn't even show which regex failed.  We've got a lot
of regexes, so it helps to report the error reasonably.

Also since we now actually capture the exception, it  means
we could in theory try to recover instead of having the whole
world blow up.