INTERIM current state is top of stack
going back to having current state be the state that's
on the top of the stack.  this idea here is that ejected
expressions won't even need to be pushed, which is more
efficient (stack slices can be clean again) but more
importantly should make it easier to do lookahead assertions
(e.g. lack-of-separator) as ejected expressions.