fpl2cc separator/comment changes
Comment and separator code are no longer separate.

Separators can be specified by code block or name.
If specified by name, we search for the corresponding
inc file in the src_path.

Space separation is still the default, but is no longer
assumed, so if you specify a separator _or_ a comment
style, space separation no longer happens.  This is prep
for supporting separator "none".

Updated comment/c.inc to integrate spaces, and changed
it so that it would eat all the spaces/comments instead
of needing to be called multiple times.