move error() out of fpl_reader; callers handle errors
Move normal syntax error reporting out of fpl_reader.
fpl_reader still supplies an error formatting function (for now),
but does not otherwise involve itself in reporting syntax errors.
It still has an on_error callback which is still used for reporting
internal errors (though that might change).

SourcePosition class has been moved to fpl_reader.h so that it
can be known to fpl2cc for purposes of error reporting.  This is
a better place for it anyway, as it directly refers to readers,
though it's probably still not ideal.  It's all an eternal work
in progress.