stringformat: '}}' evaluates to single '}'
Make '}}' evaluate to a single '}'.  This is so that
if you have a '{{' (generating a single '{'), you can
use '}}' for the closing brace, thus making braces
match in the plain text as well as the target.

this -does- change the interface:  if you actually wanted
a double closing brace, you need to write '}}}}' now.