fix Reducer::to_str
For some reason, I was using ^ as an argument separator.  Changed
to space, and now it's readable.

Also fixed comments around Reducer to explain that reducer
arguments aren't ordered.  It is confusing at the moment
that Reducer::to_str doesn't necessarily print the arguments
in the same order as they were specified in the file.
Argument order info is currently lost.  Might fix this
at one point, because it is confusing.. hmm