split fpl into separate headers; rename for filesystem
Originally, all the fpl classes were internal to fpl2cc.
The theory was that I would rapidly be at the point where
I would be rewriting it all in jest anyway, and it would
be better not to clutter up the filesystem with a bunch
of headers which were only going to be used in one temporary
place anyway.  etc etc...

Fast forward to now when I want to rewrite fpl not in jest
but in fpl.  In order to do that, I want to use the classes
I have.  This allows that. Yay.

As part of this, renamed each key class to all lowercase,
so that the class name matches the filename.  I don't love
this but I feel like it's the least-bad option, given that
some filesystems are case-insensitive.  Sigh.

Also kill spurious whitespace at end of lines.  I
accidentally did existing files as well as the new ones
when cleaning up new ones, and it seems like too much
trouble to split the existing file stuff out now.