jemplate: scons rule, vim ftdetect, #line directives
scons: jemp header creation rule accepts .h.jemp.
(which seems to still work for plain .jemp.. ok shipit)

ftdetect: Add .h.jemp and .cc.jemp to ftdetect as cpp files
so that vim highlights them as c++.  this is good enough for
the moment.

jemplate: Add #line directives to generated code so that any
error messages from code in an expansion point to the actual
sources (i.e. in the template files)