jemp-ify fpl parser class
This is essentially step 1 - the generated parser is now
generated off a jemp template, but we still have fpl_base_parser.h,
which will want to be folded in.

There's some scons hackery in here to avoid a circular fpl2cc <-> jemplpl
dependency.  In order to rebuild jemplpl now you need to:

  scons rejemp=1