fpl: Fix import rule order bug!  and other cleanups
Rules were being imported in std::set (name) order, and not
in the order they were specified in the source.  This was a
subtle disaster because order influences precedence.

Fixed and tossed in some tests and minor cleanup.

Also add the name of the state to state_string() in generated code.