jemp: whitespace controls
Text chunks are now broken by newlines and newlines with
space before or after get separate rules.  This is setup
for adding control of leading/trailing space.

Added rules for leading and trailing whitespace so we could
begin to attempt to control them.

Added "@-" and "-@" to remove leading and trailing space.
This was inspired by jinja, though this implementation is
a bit different in that it tries to be clever in dealing
with trailing space up to an indented normal text block.

I don't know if I love this implementation, but it's good
enough for now.