fpl_reader: add tests, fix bugs, make stuff public
Added tests for fpl reader.

Fixed a latent bug in char_length() wherein if the read
pointer somehow got derailed to the middle of a charater,
it wouldn't realign by itself (bad bitmask).

Fixed another latent bug where there's an expectation of an
explicit \0 in the input buffer, but no such \0 was put in
the buffer if the buffer was created from memory (as opposed
to from a file.  The \0 thing there is wonky right now
but this makes it work.

"Fixed" error handling on bad regex in fpl_reader.  This
currently includes regex compilation errors, the real fix
for which is to have callers pass in compiled regexes so
that they'll get the error at the appropriate time (in the
fpl case, at fpl compile time) as opposed to at user-runtime.
However, this fix is correct for runtime errors and moving
regex compilaton out can be done later.

Return bool from read_exact_match() instead of chucking a
pointer around.

Made a bunch of stuff which used to be private due to chucking
pointers around public, now that they are no longer chucking
pointers around.

Fix some size_t/int issues.