fix calls to debug_peek
There are 2 debug_peek calls.  One takes a length, and the other
takes a position and a length.  The one which only takes a length
defaults to starting at the current position.  The problem is
that since both position and length are (in the eye of c++) the
same type, it leads to surprising calls to the wrong one if
you call it with one arg.  So, it's boobytrappy.  Who wrote this?
Sigh.

Anyway, fixed 2 calls to debug_peek