fpl build process:  use depfiles instead of scanners
fpl2cc gets a --depfile <filename> option for making make/ninja
style dependency lines.  --depfile is either the full name
of a dependency file, or just an extension (i.e. begins with '.'),
in which case the full name is inferred from the target file.

replace fpl scanners with use of depfiles