jemplpl etc: add realign directive; use it for #line numbers
Add an @realign directive to jemp/jemplp which is like a macro
which gets inserted where jemp thinks the source and generated
lines won't match.  This lets you tell it to add stuff like #line
directives if you're generating c/c++ so that you can have
some idea where a given error actually came from in the source.

The line matching is very rough right now.  Seems like an improvement,
though, so I'm shipping it.

Also move #line within header guards in the starting #line directive
within the generated header guards so that it only sets the line number
if the generated code is compiled.

Squashed commit of:
commit 1e2878de0b31923d40c3bfb97ba1e791d7b0c851
commit bcb38a9f2cf848f5218e7c3f975afe9df4874b02