fix default construction on error case of val()
The attempt I made to return a default in the (probably unreachable)
error case of val() didn't work syntactically for pointers.  Changed
it to make a dummy variable, which I think will work for anything
which can be stored in a std::variant.

I didn't detect this because apparently there's something wrong
with the scons dependencies and this hadn't recompiled.  Not sure
where the dependencies are wrong.