fpl: @grammar etc imports comment styles if importing whole grammar
It's weird to import a "standard" grammar and not get the comments
for that grammar.  At the same time, it's weird and possibly harmful
to get all the preamble code if you're just importing a particular
construct.

So, make it such that it imports separator (comment) code if you are
importing an entire grammar, but not if importing just some subset
of that grammar.  Also, apply the same logic to preamble code - i.e.
don't import it unless importing the entire grammar.

In present practice, this changes nothing, as we're only really
importing from "pure" fpl.