fpl: support empty input
Formerly, if the top level production for an fpl was something
like "foo* -> done ;" it would fail on empty input (even though
foo* should match a lack-of-foo), and (worse) it would fail
on matching one or more foo because nothing told it to be done,
so then it would enter state 0 again and splat whatever foo it
got with an empty foo.

It occurs to me now that I'm checking this in that the crux of
this change (using more_to_parse() in parse_partial) is more
complicated than it should be, and counter to the principle of
parse_partial in that parse_partial should just parse what it
can and let callers decide if it was useful or what.  Possibly
parse_partial should go until either told to stop or it finds
a goal.  i.e. the decision should be moved up to the caller.
Checking this in and moving on for the moment, because it works.