check/report unknown scanners; lots of comment yammering
If a rule references a scanner for which there's no known
implementation, throw an error.  Without this, you get
avalanches of incomprehensible c++ errors on the generated code.

This is not implemented very efficiently, I'm realizing now that
I check it in.  It should probbaly be merged with check_actions(),
thus avoiding a second loop through the rules.  Moving on.