move default reduce to jemp
Move the default reduce code to jemp, which is good.

Default is now to try to cast/construct whatever we're producing.
This could be a good way to do things in that it could be a step
toward something more oo or data-oriented.

It might also be bad because for the 2 modern c++ compilers I know
of (gnu and clang), if there's a bad implicit cast or lack of
appropriate constructor, it dumps about a million lines of "note:"
error messages, completely obscuring the actual source of the
problem.