automatic type aliases (initial)
This might be useless, but it's the first pass at automatic type
aliasing, wherein we try to infer the type for a given production
based on the types of other productions.

The idea here is so that if you have something like:

@type_for bar int
foo -> bar ;
bar -> bat ;

.. it can guess that bat is the same type as bar and not throw
all kinds of error messages at you.