fpl2cc:  add parse_goal to generated parsers
parse_goal attempts to parse a goal from the input, but doesn't
throw errors if that's not what it got.

this is one way to do mised/partial parses.   I'm going to use this
to transition the fpl parsing from being done by the "products"
class to being done in fpl - it will allow me to do so piecemeal,
and not have to convert everything in one shot.