test of fpl construct "x? -> x ;" for defaults
Proof that (at least in some cases!) it's possible to
make an optional product reduce to itself when missing,
which lets the fpl author return a default in the reduce
action, so that rules which use the optional thing don't
have to deal with figuring out defaults.  (i.e. it moves
default evaluation to where the rest of the construct is)