stringformat columnation: no spurious space at end of line
Existing columnation trated the last column as any other - that
is to say, if the longest string in the last column was 11 chars,
every line in the last column would be padded out to 12 (11 + 1
for the separator).   Padding everything in the last col out to
11 (not 12 since the \n works as a separator) might make sense,
but I tried it and found the extra spaces at the ends of lines
annoying, especially when you're trying to compare across different
columnated outputs.

So, the new behaviour is to not pad the final column on each line
at all.