fix hex bug in c_str_escape
To my mild surprise, hex string escapes aren't limited to 2 hex
digits - they are arbitrarily long, and then converted to a char
(whatever length that is, though it's typically 8 bits nowadays).
"\x12bar" therefore gets interpreted as char 0x12ba and 'r'.

The fix was to use octal, which, for whatever reason, is limited
to 3 chars.