lack-of-separator acts like terminal
Make lack of separator act like a terminal with respect to never
preempting the next_product.

I believe this is logical and possibly necessary:  next_product
must always be the next input because we can only have (at present,
anyway) one of them, so if you start working on something else
while there's a next_product outstanding you're going to end
up splatting the next_product if you're successful parsing
whatever you're working on.

I'm noting that it's already treated as a terminal anyway in terms
of how it gets to the stack.