fpl_reader: make most constructors explicit
Several fpl_reader and utf8_buffer constructors should have
been explicit, but were not.  This led to weirdness with
strings being accidentally converted to fpl_readers, with
the reader attempting to open the file indicated by the
string.  oops.  fixed.

Also fix the error reporting function which was triggering this
- it should have been static