ansi-c.fpl: use custom scanners for enum constant etc
Due to a perceived lack of other way to do it in fpl, the
ansi-c fpl treated enumeration_constant and typedef_name
the same as identifiers.  The lex/yaccc implementation is
to return different tokens for enum constants, typedefs,
and identifiers depending on what's in the symbol table.
I realize now I could have done this already through reduce
actions, but anyway I've now hacked in scan function hooks
to do the same.

Don't take any of this fpl seriously - it's about 95% untested.
Moving on.