states is now in a map, not vector
Ever since I redid state generation to be recursive, there's
been the threat of disaster if we exceeded the preallocated
capacity of the vector, because that would invalidate the
"this" pointers of all the existsing states down the stack.

This should make it easier to remove states too, should I ever
get around to folding states or other optimizations.

I'm still keying by integer because I like to see nice, straightforward
"state 0" in dumps, etc.  much easier to read than lr_set ids.