fpl2cc: subexpression parsing; partial functionality
Subexpressions parse and work in the non-nested case.

The current implementation is to create a dummy rule
for each subexpression.

Problems are mostly around reducers for the dummy rules,
and said problems currently preclude using this in the
main way I wanted to, for parsing things like argument
lists or json objects or whatever. for example:

  '{' ( arg (','^ arg)* )? '}' -> object ;

is fail because the outer dummy rule ends up having multiple
arguments and the default reducer can't handle it.

More to come.