apply @default_action at resolve time
Apply @default_action at resolve time so that if you base something
on an abstract grammar (@grammar foo) you don't have to specify
the default reducer beforehand.  I keep abstractifying grammars
and then expecting to be able to put the @default_action in the
implementation, so clearly this is the more intuitive thing.

This does mean that you can only have one default reducer.