remove last_product*; everything is on the stack
Get rid of the basket of last_product stuff and just use the
lr_stack.  This is more traditional anyway.

Add lr_read_pos, which is the index in the stack from which
the next nonterminal (if any) is to be read.  In theory, we
could stack up multiple items on the lr_stack before parsing
them, but at the moment the lr_read_pos is always either
the top item in the stack or one past (= nothing), except
during reduce, where it points to the items to be consumed.

lr_frame_start was redundant with lr_read_pos - removed.