stop using fprintf in fpl error handling
In a vain effort to plug a recent heisenbug wherein random bytes
appear to be getting swapped in warning output, convert fpl
error message handlers to use std::cerr stuff instead of fprintf.

I was using fprintf before reinventing stringformat, but now that
that's been reinvented, fprintf isn't really helpful, and I was
wondering if the problem was something screwy in the normal clib
buffering.  However, the problem persists, so it's not that.