scons run_and_capture_action: show signal based failures
Prior to this, if a comand failed due to (eg) a segv or interrupt,
there was no particular message shown to whoever's running the
build.  The user would have to know to check the capture file
to get the return code and then interpret that somehow.

This change adds a (hopefully helpful) message on such cases.