fpl imports:  path fixes and other improvements
Search relative to source -after- searching the fpl
import path (src_path).  Not only does this make stuff
actually work, but on rethink it makes more sense anyway.
Users can always change the command line if they want to
import from someplace else.

fpl options: don't always add '.' to the front.  What we
want is to add the path relative to the script, but '.'
is the current working directory.  Not the same at all.

Fixes to tests.

Minor cleanup.