use asyncio instead of threads on interactive run_and_capture
Because the subthreads which were supposed to read stdout and stderr
would remain blocked on a read at the end of a process, the whole
thing was cludged to make it so that the parent process would
read anything remaining in the output streams after subprocess
termination.  But, that creates a race between the parent process
and the reader threads.  But, without the 2-place read there's
a race between the end of the process and the reads.

There were other possible fixes. Long story short, this is converted
to asyncio and now seems to actually dtrt.  Moving on.