fix run_and_capture_action (build issue)
Fix the run_and_capture_action function to take a varlist,
and to return an actual action instead of just a command
string, which makes it so that scons can tell it needs to
recompile if things like INTERACTIVE or FPLOPTS changed.

The problem was that without the varlist, scons doesn't know
to update the build signature when (eg) FPLOPTS changed.

Come to think of it, if I were smart, I'd args/kwargs it up and
have it pass all other arguments on to the Action factory. hmm.
maybe next time.