make to_hex be reasonable for ints; tests
The old to_hex basically just dumped the bytes of whatever it
was given, starting at the lowest address.  This is (or can be)
wrong (or at least counter intuitive, and therefore probably
wrong) for several cases, such as where the word size is less
than the size of the int.  For example, on my particular machine,
if I have a long long = 0x2300000000000017, it actually gets
printed as 0000001700000023, which is how it is stored, but
not what I think I or anyone would expect.

So, this makes an exception in to_hex() for all integer types
larger than a char.

Probably what I actually want is a separate function for dumping
bytes vs dumping values.. I dunno.  moving on for now.