json.fpl and test: make it less of a mess
fpl json.test: move to text format, simplify, use more recent
fpl features.

It was too hard to read the diffs in the binary format, and that
was hiding a bug (which turned out to be a bug in the test, and
not fpl, but the point remains.

Move the separator code to grammarlib/json.fpl because that's
where it belongs.  It's annoying that the separator code is
embedded c but that's more of an asthetic problem than a real one.