Check copyableness of product types
If you try to add a non-copy-construtable type to a std::variant,
it causes a completely un-humam-readable explosion of error messages,
none of which seem to tell you what actually caused the problem
(though maybe they do, someplace burried in 200000 lines of template
expansions).

Adding a static assert reduces the blather considerably (though not
enough, imo).  It's reduced enough that it's actually possible to
determine from the error message what the problem was, though,
so shipit.