Fix 2 build dependency issues (fpl/jgt)
Include subgrammar imports in .deps file

This was foolishly only including the top level imports in the .deps
files, which obviously lead to scons not understanding that it needed
to recompile when a file 2 levels (or more) down changed.

Also use the correct dependency emitter in the grammar test Jgt rule.