fpl: allow passing product to production_rule()
The only reason this didn't already take a product name is that
when parsing fpl (presently) we don't know the product (yet) at
the time when we create the rule.

This may change in the future, and I'm going to add a case now
where the product name _is_ known at time of creation, so I
changed the constructor to be (I think) slightly more correct
in this regard.