fpl: fix issue 18
Add an lr_transition type for "COMPLETION", which denotes a transition
from parsing to being done with parsing.

Change how goals are handled:  instead of giving the set of goals to
the state generator, create a completion rule for each goal (each
with target "_fpl_goal"), and give _fpl_goal to the state generator.

This fixes issue 18, which was ultimately a termination issue.  Formerly,
we didn't temrinate on matching a goal because goals could be part
of other goals, etc.  But adding (an) explicit termination rule(s)
disambiguates termination.

Also debug and comment updates.