fpl: get rid of "FAILULAR" error on val() mismatch
Remove the error on calls to val() because the only (?) time it gets
hit is when the parser has a different result from the output type
due to parsing stopping due to syntax errors or such.

val() still needs to go.  I just don't have a bette solution off the
top of my head right now.

Possibly nothing cares about the return value from parse() anyway.
Ideally, I'd like it to return a value, but it might not be practical
to do so now that different parse results can have different types.
Callers can get the final result from result() in any case.