fpl: make it so that TERM_EXACTs don't mask each other
So, I recently hit a "doh" issue where, in the jest grammar,
if you have a construct like foo += bar;, instead of parsing
as 'foo', '+=', 'bar', it parses as 'foo', '+', "HEY YOU CAN'T
ADD AN '='" because (for various reasons) the scan for '+'
was before the scan for "+=".

This was obviously horrible, and I'm not sure if there was a
way other than using regexes or custom terminals, which seems
nuts.  So, this essentially implements longest match (for
exact match terminals only).

Not that this change -doesn't- check anything with respect to
regex or custom terminals.  They can still mask each other,
undetected.