rejemp; fix @@ in jemptest; dump states on rejemp
There was a '@@'string in jemptest which, prior to the change
making TERM_EXACT stop masking each other, would parse as '@'
to end a subst and then another '@' to start one.  Now that it
does longest match, '@@' was being matched.  So I changed the
test.

Also checking in the state dump as well so that it can be used
as evidence in case of breakage.  Hoping it won't cause merge
issues.