fpl: handle more ambiguous grammars
Handle ambiguous grammars by marking not-ultimately-matching
symbols on the stack and unwiding such that the parser can
try other possibilities on an ambiguity.

I'm slightly surprised that this can handle 3- (or more) way
ambiguities, but it seems to, and anyway I believe it will
solve the case I want solved, so I'm shipping it.

BTW it's not terribly efficient.  There's lots of room for
optimization, should we care.

Squashed commit highlights:

commit 9610f13f4b886e328728603299b420d67016d295 fpl: make nonterminal mismatches dtrt
commit 733883de9bdcb93b59ccd1bf9790e2ccab9f9989 fpl: "lookahead"/rewinds work at all