fpl: --dump-states dumps explicit transitions as well
State dumps currently show all the lr_items for the state, but
these are not necessarily in the order they'll be used - the
set of transitions for the state shows that.  But, the transitions
are (currently) ugly and don't give the whole picture. So, for the
moment, show them both on state dumps.

Also cache generated transitions so they aren't regenerated.  In
theory, the cache needs to be cleared on state changes, but at the
moment that doesn't happen so I'm punting that.