jemp: trailing whitespace can include multiple \n
If the trailing whitespace is leading to another @ construct
(substritution etc), match multiple \n to fold empty lines
leading to the next construct.

The idea of this version of trailing whitespace is that if you
indent @ constructs for readability in the jemp source, it
can end up being superflous whitespace in the gnerated text.
Folding multiple blank lines in thic ase is in line with the
original intent.