fpl: make custom scanners able to return bool
This involves some c++ horror, because apparently you can't
if constexpr() outside a template (if you do, it'll evaluate
constexprs as if they weren't constexpres) and you can't
"else" an if constexpr() anywhere (or, you can, but it will
do the else even if the constexpr(s) evaluate to true).

The first problem (no constexpr outside template) prevents this
from doing reasonable... well, anything reasonable, including
getting reasonable line numbers in errors.

The second problem means we need to duplicate the checks
in the if constexpr/else if constexpr stuff in a sepaprate
stateic_assert to get a reasonable error message if there's
a return value of a type we don't handle in the scanner.

Maybe some template matching magic could get this structured better
but I'm done dealing with c++