fpl: custom scanners return Terminals
Get rid of the term_custom_result monstrosity in favor of
having custom scanners return Terminals.

Terminals are the obvious thing - the whole point of them
is to encapsulate a match or lack of match.

The only reason for the term_custom_result stuff was that
I failed an int roll.