fpl 1.2:  @import, `import`, @grammar, @goal changes
Refactor imports such that they are 80% sane.

Make the import grammar more consistent.  This means if you try to
import a goal, you need to make it (more) explicit by enclosing
the grammar name in backticks.

@grammar no longer supports pulling in a subset of another grammar.

Lots of yammering about the different cases.  This could still
be made more sane and consistent but I'm moving on for now.