fpl: no-op change to subex_varname() call
Pass the expression from the step instead of the raw expression to
subex_varname.

This is part of another refactor I'm doing on another branch.  The
point is that it amkes it clear that the step is actually self
contained and we don't have to pass the raw expression string
around with it.