fpl: fix repetition interactions
This change makes rules like:

  'p'*:a 'p'?:b 'p':c -> wup;

work such that parameter "c" is always passed, "b" and c" are passed
if there are 2 p's in a row, all are passed if there are more than
that (as opposed to "a" eating all the arguments).

This is either a fix or a feature, depending on how you look at it.