site_scons monkeypatch waitpid/add on-exit profiling
Rework site_scons/site_init.py to allow gathering of profiling
data on spawned processes.

This implementation catches python-level waitpid() calls, and, if
it's available, calls wait4() instead to obtain the cpu/memory
usage of every subproc.  I did it this way because the exit callbacks
in asyncio all happen after the wait() calls, at which point it's
too late to get the profiling data.  It's not my favorite implementation
but it's the only way I can think of to reliably get usage info
upon exit, without racing.