Check in profiling data under yprof; preserve tree
Rearrange profiling output so that the tree is preserved,
and set the top level profiling output directory to "yprof".

Why y?  Because I want profiling data to sort later in directory
listings, and zprof is already a thing.

In the yprof dir, .gitattributes is set to "union" merge these,
and since profiling output is line-oriented, I expect no merge
madness.  It should end up just being a nice record.