wcalc:  tweak integer parsing and add a comment
I looked at wcalc.fpl, and was temporarily confused at how it
magically handled operations on mixed types (integers and doubles).
Of course it doesn't, but this lead me to (1) change the integer
reduce to actually parse an integer instead of parsing a double
and implicitly converting it to an integer, and (2) add a comment
explaining.