fpl: modify single quote string escape rules
The existing single quote escaping didn't make sense.  For example,
if you did '\\' you'd actually get a string with 2 backslashes.
Very unintuitive.  Now, the way it works is the backslash only
applies to single quote and itself.

So, '\\' is a single backslash, '\'' is a single quote, but (eg)
'\x23' is a 4 char literal (including a backslash).