fpl_reader: equality operator for SourcePosition
You can now compare source positions for equality.

I'd like to be able to compare them quantitatively
(< > etc), but that's not trivial since the positions
could be in separate files, and ideally such a comparison
would be based on which file was opened (or read) first,
but that doesn't look practical to implement right now.
Moving on.