fpl2cc scons:  "always" remake .h.jemp
jemplpl doesn't yet generate dependencies in any form which
we could use from scons (or anyhting else).  This means that
if a .jemp file includes another one, scons doesn't know to
recompile the top level .jemp.

So, as a stopgap to avoid forgetting to recompile .jemp files,
tell scons to always rebuild them.