fpl: default main now prints output
Default main prints output now, because that's what nearly every
test/existing case which uses it wants (in some form or other).

Many test cases changed to match this.

There's a bit of awkwardness in the wcalc test around numeric
vs string types, which I punted on by making it just stick to
numeric.  But, it would be nice to be able to fix it back to
how it was (which was printing "result: xxx" instead of just
"xxx").