run_cc_tests: test output depends on test source
Make cc test output depend on test source, as well as on the generated
test program.

Doing this because I sometimes make tests use their own source code for
input data, and I'm happy to brute-force rerun any tests whose source
code has changed anyway.  Rerunning on a source change (even if it's,
for example, a comment) is what I always seem to expect (for tests,
anyway).

The cc tests are fast so it's not much cost, and better to err on the
side of actually running the tests.