grammarlib/json: don't include quotes on strings
The grammarlib json fpl was written so as to (by default)
include the quotes on the return value for a string match.
This is rarely what anyone wants, so I changed it to just
include the guts.

The wierd part (still) is that strings returned by default
will include the backslashes on any escaped quotes or whatever
within the string.  I'm not yet sure how to handle that.