src/jemplpl/.gitattributes:  merge strategy "ours" on jemplpl
This also required:

   git config --global merge.ours.driver true

which tells it to use the "true" command to merge these files.

git docs seem to imply that there should already be an "ours"
strategy, but without the above command it seems to silently
use the default strategy.  I'm not sure how to propagate the
git-config up, also.  Maybe the config can be put in a
.gitattributes as well?  Moving on, for the moment.

Here's where I got this plan:
https://stackoverflow.com/questions/928646/have-git-select-local-version-on-merge-conflict-on-a-specific-file