site_scons:  add conf_command()
Add a python function for getting config from shell(ish) commands.

This is so that we can safely glean configuration from commands
like llvm-config, even if such commands aren't installed.

Slightly overengineered in that is supports pipes.  Does what
I think is correct w/ respect to shell variables - that is,
they are passed through to be expanded when scons actions are
run (not expanded at config time).  Maybe unerengineered in
other ways.  Ship it.