fpl2cc:  allow stub @scanners
@scanners can be specified without an implementation, in which case
a stub never-matching scanner is filled in and a warning is given
if/when such a scanner is used in a rule.

This allows authors of abstract grammars to specify any necessary
scanners, without having to fill in an implemenation while also
automatically informing useds of the abstract grammar about what
they need to implement.