fpl: add @embed for embedding header files
Add @embed to fpl, which will embed a c++ header (or whatever
else you want) in the generated code.

The idea was to make it so that generated code could be compiled,
regardless of whether or not whoever's compiling it has access
to (eg) the headers in the util dir.  However, it turns out I
really need to add this feature to jemp and not fpl.  But,
I wrote it, it works, so I'm checking it in in case it comes
in handy.