fpl: add @default_type
This lets you set the default type separately from the output
type.  Also, if the default type is set, it preceeds type
inheritance.

This is, at this point, too many features.  But, I think the
thing to do is to keep this but remove one or more existing
type related features.