enable yprof on main;  change column order
Save profiling data in tsv files in the "yprof" directory when
building on the main branch.  These files relate performance
data (from tests etc) to the current commit ID.

This does (presently) assume that if someone is building on the
main branch, it's with a clean workspace, which is (for all intents
and purposes) the case for the workflow on this project.

Incidentally, I changed change the column order and some formatting
from how it was on the jest2 branch.  I was unhappy with how I ordered
them there, but didn't want to remove or mess with existing profiling
data on that project.  Since this is a fresh start, I don't have to
change anything retroactively, so I've changed it.  Let's see if I
actually hate this less, though.